omaplfb: If the framebuffer is in VRAM avoid mapping into Tiler1D
Calculate whether we get an address which lies in the range of the frame
buffer, if we do pass this on directly to dsscomp.

Previously the framebuffer was accidentally being mapped into Tiler1D,
which could cause some composition scenarios to fail to render because
HWC calculations assumed exclusive use of the tiler 1d slot for itself.

Change-Id: I7c90e4adb9bee25e6a3825d456653af088803f83
Signed-off-by: Tony Lofthouse <a0741364@ti.com>